+++ CHANGES.rst
@@ -1,6 +1,8 @@
 0.9.0 (in development)
 ======================
 
+* Added ``form.getData()`` to get input data more conveniently by field name in
+  prefixed forms, such as forms in a ``FormSet``.
 * ``clean()`` functions can now specify which fields they make use of for
   cross-field validation. When provided, ``clean()`` will only be called during
   partial form updates (e.g. ``onChange`` handling) if one of the specified